+++ src/Telegram.Bot/Types/ChatPhoto.cs
@@ -16,9 +16,21 @@ namespace Telegram.Bot.Types
         public string BigFileId { get; set; }
 
         /// <summary>
+        /// Unique file identifier of big (640x640) chat photo, which is supposed to be the same over time and for different bots. Can't be used to download or reuse the file.
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public string BigFileUniqueId { get; set; }
+
+        /// <summary>
         /// File id of the small version of this <see cref="ChatPhoto"/>
         /// </summary>
         [JsonProperty(Required = Required.Always)]
         public string SmallFileId { get; set; }
+
+        /// <summary>
+        /// Unique file identifier of small (160x160) chat photo, which is supposed to be the same over time and for different bots. Can't be used to download or reuse the file.
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public string SmallFileUniqueId { get; set; }
     }
 }