+++ src/Telegram.Bot/Types/FileBase.cs
@@ -10,10 +10,16 @@ namespace Telegram.Bot.Types
     public abstract class FileBase
     {
         /// <summary>
-        /// Unique identifier for this file
+        /// Identifier for this file, which can be used to download or reuse the file
         /// </summary>
         [JsonProperty(Required = Required.Always)]
         public string FileId { get; set; }
+
+        /// <summary>
+        /// Unique identifier for this file, which is supposed to be the same over time and for different bots. Can't be used to download or reuse the file.
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public string FileUniqueId { get; set; }
 
         /// <summary>
         /// Optional. File size, if known